Post successful whitelist additions to slack
* Moved slack code to slackutils.go
* Add posting message to slack channel when whitelist entry saved
* Added logging whitelist entry submissions
